move the format_message function & stop_all_event_loop function into a processor utils module
