+++ yupana/processor/abstract_processor.py
@@ -26,10 +26,11 @@ from enum import Enum
 
 import pytz
 from django.db import transaction
+from processor.processor_utils import (format_message,
+                                       stop_all_event_loops)
 from processor.report_consumer import (DB_ERRORS,
                                        QPCReportException,
-                                       format_message,
-                                       stop_all_event_loops)
+                                       ReportConsumer)
 from prometheus_client import Counter, Gauge, Summary
 
 from api.models import (Report,
@@ -148,7 +149,7 @@ class AbstractProcessor(ABC):  # pylint: disable=too-many-instance-attributes
                 try:
                     self.assign_object()
                 except Exception:  # pylint:disable=broad-except
-                    stop_all_event_loops()
+                    stop_all_event_loops(ReportConsumer)
             if self.report_or_slice:
                 try:
                     await self.delegate_state()
@@ -418,7 +419,7 @@ class AbstractProcessor(ABC):  # pylint: disable=too-many-instance-attributes
                 'Could not update %s record due to the following error %s.' % (
                     self.object_prefix.lower(), str(error)),
                 account_number=self.account_number, report_platform_id=self.report_platform_id))
-            stop_all_event_loops()
+            stop_all_event_loops(ReportConsumer)
 
     def move_candidates_to_failed(self):
         """Before entering a failed state any candidates should be moved to the failed hosts."""
